refactor(plugin-settings): extract DefaultSettings to components
Move the generic schema-driven settings surface out of the surface
module into its own #components entry, mirroring the components/
convention used by other plugins.
