+++ packages/plugins/plugin-settings/package.json
@@ -14,6 +14,7 @@
   "type": "module",
   "imports": {
     "#capabilities": "./src/capabilities/index.ts",
+    "#components": "./src/components/index.ts",
     "#meta": "./src/meta.ts",
     "#operations": "./src/operations/index.ts",
     "#plugin": "./src/SettingsPlugin.ts",
