+++ packages/plugins/plugin-settings/src/components/DefaultSettings/index.ts
@@ -1,0 +1,5 @@
+//
+// Copyright 2025 DXOS.org
+//
+
+export * from './DefaultSettings';
